+++ packages/w3p/src/helpers/eth.ts
@@ -1,6 +1,5 @@
 import { DECIMALS } from '@distributedlab/tools'
 import { type providers, utils } from 'ethers'
-import cloneDeep from 'lodash/cloneDeep'
 
 import { EIP1193, EIP1474 } from '@/enums'
 import { errors } from '@/errors'
@@ -53,7 +52,10 @@ export const connectEthAccounts = async (provider: providers.Web3Provider) => {
   await provider.send('eth_requestAccounts', [])
 }
 
-export function handleEthError(error: EthProviderRpcError): never {
+export function handleEthError(
+  error: EthProviderRpcError,
+  defaultHandlerFn?: (error: EthProviderRpcError) => unknown,
+): never {
   switch (error.code) {
     case EIP1193.UserRejectedRequest:
       throw new errors.ProviderUserRejectedRequest(error)
@@ -92,24 +94,34 @@ export function handleEthError(error: EthProviderRpcError): never {
     case EIP1474.JsonRpcVersionNotSupported:
       throw new errors.ProviderJsonRpcVersionNotSupported(error)
     default:
+      if (defaultHandlerFn) {
+        defaultHandlerFn(error)
+      }
+
       throw error
   }
 }
 
-export const wrapExternalProvider = (provider: providers.ExternalProvider) => {
-  const p = cloneDeep(provider)
+export const wrapExternalEthProvider = (
+  provider: providers.ExternalProvider,
+  errorHandler?: (error: Error) => unknown,
+) => {
+  const _baseRequest = provider.request?.bind(provider)
 
-  p.request = async (request: { method: string; params?: Array<unknown> }) => {
+  provider.request = async (request: {
+    method: string
+    params?: Array<unknown>
+  }) => {
     let result: unknown
 
     try {
-      result = await provider?.request?.(request)
+      result = await _baseRequest?.(request)
     } catch (error) {
-      handleEthError(error as EthProviderRpcError)
+      handleEthError(error as EthProviderRpcError, errorHandler)
     }
 
     return result
   }
 
-  return p
+  return provider
 }