+++ packages/w3p/src/types/ethereum.ts
@@ -16,7 +16,7 @@ export interface EthereumProvider extends providers.ExternalProvider {
   removeAllListeners(event?: string | symbol): this
   providers?: this[]
   selectedAddress: string | null
-  request: (...args: any[]) => Promise<void>
+  request?: (request: { method: string; params?: Array<any> }) => Promise<any>
 }
 
 export interface EthProviderRpcError extends Error {