+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0-rc.14] - 2023-05-18
+### Added
+- `@distributedlab/w3p`
+  - `ProviderDetector` - configuration and custom error handlers passing
+  - `handleEthError` - optional custom error handlers
+
+### Fixed
+- `@distributedlab/w3p` - `wrapExternalEthProvider` not supported provider listeners
+
 ## [0.2.0-rc.13] - 2023-05-18
 ### Fixed
 - `@distributedlab/fetcher` - error building in response
@@ -207,7 +216,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [old repo]: https://github.com/distributed-lab/web-kit-old
 
-[Unreleased]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.13...HEAD
+[Unreleased]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.14...HEAD
+[0.2.0-rc.14]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.13...0.2.0-rc.14
 [0.2.0-rc.13]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.12...0.2.0-rc.13
 [0.2.0-rc.12]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.11...0.2.0-rc.12
 [0.2.0-rc.11]: https://github.com/distributed-lab/web-kit/compare/0.2.0-rc.10...0.2.0-rc.11